requirements.txt
fastapi
pdfplumber
python-dotenv
google-generativeai
uuid
uvicorn
python-multipart
python-docx
reportlab
dotenv

